refactor: moved emotion and constants to separate files
