+++ src/components/points-list-information/constants.ts
@@ -1,0 +1,1 @@
+export const cols = ['Пункт', 'Адрес', 'Город'];
